reorganizing code
