+++ joblib/compressor.py
@@ -2,8 +2,9 @@
 
 import io
 import zlib
+from distutils.version import LooseVersion
 
-from ._compat import _basestring
+from ._compat import _basestring, PY3_OR_LATER
 
 try:
     from threading import RLock
@@ -26,27 +27,6 @@ _PREFIXES_MAX_LEN = max(len(prefix) for prefix in (_ZFILE_PREFIX, _ZLIB_PREFIX,
                                                    _GZIP_PREFIX, _BZ2_PREFIX,
                                                    _XZ_PREFIX, _LZMA_PREFIX,
                                                    _LZ4_PREFIX))
-
-
-class CompressorWrapper():
-    """A wrapper around a compressor file object."""
-
-    def __init__(self, obj, prefix='', extension=''):
-        self.obj = obj
-        self.prefix = prefix
-        self.extension = extension
-
-    def compressor_file(self, fileobj, compresslevel=None):
-        """Returns an instance of a compressor file object."""
-        if compresslevel is None:
-            return self.obj(fileobj, 'wb')
-        else:
-            return self.obj(fileobj, 'wb', compresslevel=compresslevel)
-
-    def decompressor_file(self, fileobj):
-        """Returns an instance of a decompressor file object."""
-        return self.obj(fileobj, 'rb')
-
 
 ###############################################################################
 #  base file compression/decompression object definition
@@ -384,6 +364,170 @@ class BinaryGzipFile(BinaryZlibFile):
     wbits = 31  # zlib compressor/decompressor wbits value for gzip format.
 
 
+class CompressorWrapper():
+    """A wrapper around a compressor file object."""
+
+    def __init__(self, obj, prefix='', extension=''):
+        self.obj = obj
+        self.prefix = prefix
+        self.extension = extension
+
+    def compressor_file(self, fileobj, compresslevel=None):
+        """Returns an instance of a compressor file object."""
+        if compresslevel is None:
+            return self.obj(fileobj, 'wb')
+        else:
+            return self.obj(fileobj, 'wb', compresslevel=compresslevel)
+
+    def decompressor_file(self, fileobj):
+        """Returns an instance of a decompressor file object."""
+        return self.obj(fileobj, 'rb')
+
+
+try:
+    import lzma
+except ImportError:
+    lzma = None
+
+
+class LZMACompressorWrapper(CompressorWrapper):
+
+    def __init__(self):
+        self.prefix = _LZMA_PREFIX
+        self.extension = '.lzma'
+        if lzma is not None:
+            self.obj = lzma.LZMAFile
+        else:
+            self.obj = None
+
+    def compressor_file(self, fileobj, compresslevel=None):
+        """Returns an instance of a compressor file object."""
+        if compresslevel is None:
+            return self.obj(fileobj, 'wb', format=lzma.FORMAT_ALONE)
+        else:
+            return self.obj(fileobj, 'wb', format=lzma.FORMAT_ALONE,
+                            preset=compresslevel)
+
+    def decompressor_file(self, fileobj):
+        """Returns an instance of a decompressor file object."""
+        if PY3_OR_LATER and lzma is not None:
+            # We support lzma only in python 3 because in python 2 users
+            # may have installed the pyliblzma package, which also provides
+            # the lzma module, but that unfortunately doesn't fully support
+            # the buffer interface required by joblib.
+            # See https://github.com/joblib/joblib/issues/403 for details.
+            return lzma.LZMAFile(fileobj, 'rb')
+        else:
+            raise NotImplementedError("Lzma decompression is not "
+                                      "supported for this version of "
+                                      "python ({}.{})"
+                                      .format(sys.version_info[0],
+                                              sys.version_info[1]))
+
+
+class XZCompressorWrapper(LZMACompressorWrapper):
+
+    def __init__(self):
+        self.prefix = _XZ_PREFIX
+        self.extension = '.xz'
+        if lzma is not None:
+            self.obj = lzma.LZMAFile
+        else:
+            self.obj = None
+
+    def compressor_file(self, fileobj, compresslevel=None):
+        """Returns an instance of a compressor file object."""
+        if compresslevel is None:
+            return self.obj(fileobj, 'wb', check=lzma.CHECK_NONE)
+        else:
+            return self.obj(fileobj, 'wb', check=lzma.CHECK_NONE,
+                            preset=compresslevel)
+
+
+try:
+    import bz2
+except ImportError:
+    bz2 = None
+
+
+class BZ2CompressorWrapper(CompressorWrapper):
+
+    def __init__(self):
+        self.prefix = _BZ2_PREFIX
+        self.extension = '.bz2'
+        if bz2 is not None:
+            self.obj = bz2.BZ2File
+        else:
+            self.obj = None
+
+    def _check_versions(self):
+        if bz2 is None:
+            raise ValueError('bz2 module is not compiled on your python '
+                             'standard library.')
+
+    def compressor_file(self, fileobj, compresslevel=None):
+        """Returns an instance of a compressor file object."""
+        self._check_versions()
+        if compresslevel is None:
+            return self.obj(fileobj, 'wb')
+        else:
+            return self.obj(fileobj, 'wb', compresslevel=compresslevel)
+
+    def decompressor_file(self, fileobj):
+        """Returns an instance of a decompressor file object."""
+        self._check_versions()
+        if PY3_OR_LATER:
+            fileobj = self.obj(fileobj, 'rb')
+        else:
+            # In python 2, BZ2File doesn't support a fileobj opened in
+            # binary mode. In this case, we pass the filename.
+            fileobj = self.obj(fileobj.name, 'rb')
+        return fileobj
+
+
+try:
+    import lz4
+    if PY3_OR_LATER:
+        import lz4.frame
+except ImportError:
+    lz4 = None
+
+LZ4_NOT_INSTALLED_ERROR = ('LZ4 in not installed. Install it with pip: '
+                           'http://python-lz4.readthedocs.io/')
+
+
+class LZ4CompressorWrapper(CompressorWrapper):
+
+    def __init__(self):
+        self.prefix = _LZ4_PREFIX
+        self.extension = '.lz4'
+        if PY3_OR_LATER and lz4 is not None:
+            self.obj = lz4.frame.LZ4FrameFile
+        else:
+            self.obj = None
+
+    def _check_versions(self):
+        if not PY3_OR_LATER:
+            raise ValueError('lz4 compression is only available with '
+                             'python < 3.')
+
+        if lz4 is None or LooseVersion(lz4.__version__) < LooseVersion('0.19'):
+            raise ValueError(LZ4_NOT_INSTALLED_ERROR)
+
+    def compressor_file(self, fileobj, compresslevel=None):
+        """Returns an instance of a compressor file object."""
+        self._check_versions()
+        if compresslevel is None:
+            return self.obj(fileobj, 'wb')
+        else:
+            return self.obj(fileobj, 'wb', compression_level=compresslevel)
+
+    def decompressor_file(self, fileobj):
+        """Returns an instance of a decompressor file object."""
+        self._check_versions()
+        return self.obj(fileobj, 'rb')
+
+
 def register_compressor(compressor_name, compressor,
                         force=False):
     """Register a compressor implementing the Python file object interface."""
@@ -392,9 +536,7 @@ def register_compressor(compressor_name, compressor,
         raise ValueError("Compressor name should be a string, "
                          "'{}' given.".format(compressor_name))
 
-    if (not hasattr(compressor, 'obj') or
-            not hasattr(compressor, 'prefix') or
-            not hasattr(compressor, 'extension')):
+    if not isinstance(compressor, CompressorWrapper):
         raise ValueError("Compressor should implement the CompressorWrapper "
                          "interface, '{}' given.".format(compressor))
